BlockFetch client: isolate impl of `BlockFetchConsensusInterface`
